Relation AI
A simple AI which finds relation between two numbers
This is the simplest form of AI , in which you give one number as a input with output.
You have to give lots of inputs with output. Your program finds the relation between input and output. 
Using that relation, you can use it to calculate the output of a different input.