The NeuralNet
The most interesting things to note in the NeuralNet class are the m_learningRate member variable, the Train() methods, the BackPropogation() method, and the Initialize() method.
The Train() methods, just applies an input, pulses the net, and performs the back propagation necessary for learning.
The meat of the network is the BackPropogation() method. This is where the bulk of the processing takes place. 
After BackPropogation apply the lessons learned by comparing the actual vs. expected output. 
Just call ApplyLearning() and the neural net will be updated.